falta arreglar los ultimos 6 errores!!!

+++ proyecto-integrador/controller/postController.js
@@ -1,5 +1,5 @@
 const db = require("../db/models")
-// const post = db.post;
+//  const post = db.post;
 
 let postController = {
     agregar: function (req,res) {
@@ -117,22 +117,86 @@ let postController = {
     },
 
     borrar: function(req, res) {
-        db.post.destroy({
-            where: {
-                id: req.params.id
-            }
-        })
-        res.redirect('/user/miPerfil')
+        let usuario_id = req.session.usuarioLog.id
+        let id_post = req.params.id;
 
 
+        db.post.findByPk(id_post)
+
+        .then(function(post){
+        if (usuario_id != undefined && usuario_id == post.usuario_id) {
+           
+            
+            db.post.destroy({
+                where: {
+                    id: id_post
+                }
+            })
+    
+            .then(function(){
+    
+                res.redirect("/user/miPerfil")
+            })  
+       
+        }else {
+            res.redirect("/user/home")
+        }})
     },
 
+    
+
+
     editar: function(req, res) {
-        db.post.findAll(req.params.id)
-        .then(function(post) {
-            res.render('edit')
+        db.post.findByPk(req.params.id)
+        .then(function(postAEditar) {
+            res.render('editarPost', {postAEditar: postAEditar})
         })
+    },
+
+    editarPost: function(req,res){
+        let id_post = req.params.id;
+        let usuario_id = req.session.usuarioLog.id;
+        let url_perfil = req.body.url_perfil;
+        let texto_post = req.body.texto_post
+
+        let post = {
+            usuario_id: usuario_id,
+            url_perfil : url_perfil,
+            texto_post: texto_post
+        }
+
+        console.log('000000000000000000000');
+
+        db.post.findByPk(id_post)
+
+        .then(function(postAEditar){
+            console.log(postAEditar);
+            if (usuario_id != null && usuario_id == postAEditar.usuario_id) {
+           
+            
+                db.post.update({ 
+                    texto_post : post.texto_post,
+                    url_perfil : post.url_perfil
+                    
+                },
+                {
+                    where: {
+                        id : id_post
+                    }
+                })
+        
+                .then(function(){
+        
+                    res.redirect("/post/detalle/" + id_post)
+                })  
+           
+            } else {
+                res.redirect("/post/home")
+            } 
+        })
+      
     }
+
